Updated menus
I changed the menus at the top so that the Publication and the Posts
links go to the main Publication page and Posts page, respectively,
instead of a widget on the home page. For now, Projects still points to
the widget.

Also, I changed the look of the main Posts page, /post/.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   config/_default/menus.yaml
	modified:   content/post/_index.md